[TASK] preprocessing for simple storage in separate submodule
